dope-backend/requirements.txt
uvicorn~=0.27.0
aiohttp~=3.9.1
botocore==1.34.26
fastapi~=0.109.0
boto3==1.34.26
python-dotenv~=1.0.1
python-multipart
pydantic~=2.5.3